tx_tools/__init__.py
import os
from dotenv import load_dotenv, find_dotenv
load_dotenv(find_dotenv())
etherscan_api_key = os.getenv('ETHERSCAN_API_KEY')


from .get_abi import get_abi
from .get_calldata import get_calldata
from .get_transaction import get_transaction

from .keccak import (
    keccak_to_hex,
    keccak_sig_to_selector,
    )


__all__ = [
    'get_abi',
    'get_calldata',
    'get_transaction',
    'etherscan_api_key',
    'keccak_to_hex',
    'keccak_sig_to_selector',
    ]
